split robot into own file/split out limits/introduce dataclass
